@@ -1,7 +1,8 @@
 {
-  "name": "design-system",
+  "name": "lab-ds",
   "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -13,10 +14,11 @@
   "dependencies": {
     "@radix-ui/react-checkbox": "^1.0.0",
     "@radix-ui/react-slot": "^1.0.0",
+    "@storybook/addon-a11y": "^6.5.12",
     "clsx": "^1.2.1",
     "phosphor-react": "^1.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
@@ -28,14 +30,14 @@
     "@storybook/react": "^6.5.12",
     "@storybook/storybook-deployer": "^2.8.12",
     "@storybook/testing-library": "^0.0.13",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.1.0",
     "autoprefixer": "^10.4.12",
     "babel-loader": "^8.2.5",
     "postcss": "^8.4.17",
     "tailwindcss": "^3.1.8",
-    "typescript": "^4.5.4",
-    "vite": "^3.1.7"
+    "typescript": "^4.6.4",
+    "vite": "^3.1.0"
   }
 }
