@@ -6,7 +6,8 @@ module.exports = {
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@storybook/addon-interactions"
+    "@storybook/addon-interactions",
+    '@storybook/addon-a11y'
   ],
   "framework": "@storybook/react",
   "core": {
@@ -15,9 +16,9 @@ module.exports = {
   "features": {
     "storyStoreV7": true
   },
-  viteFinale: (config, { configType }) => {
+  viteFinal: (config, { configType }) => {
     if (configType === 'PRODUCTION') {
-      config.base = '/design-system'
+      config.base = '/ignite-lab-design-system'
     }
 
     return config
