@@ -2,5 +2,5 @@ import { addons } from '@storybook/addons'
 import { themes } from '@storybook/theming'
 
 addons.setConfig({
-  theme: themes.dark
+  theme: themes.dark,
 })
