@@ -1,6 +1,6 @@
 import { themes } from '@storybook/theming'
 
-import '../src/styles/global.css'
+import '../src/styles/global.css';
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
@@ -12,5 +12,5 @@ export const parameters = {
   },
   docs: {
     theme: themes.dark,
-  }
+  },
 }
