@@ -1,22 +1,24 @@
+import { ButtonHTMLAttributes, ReactNode } from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { clsx } from 'clsx';
-import { ReactNode } from 'react';
 
-export interface IButtonProps {
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   children: ReactNode;
   asChild?: boolean;
 }
 
-export function Button({ children, asChild }: IButtonProps) {
-  const Comp = asChild ? Slot : 'button'
+export function Button({ children, asChild, className, ...props }: ButtonProps) {
+  const Comp = asChild ? Slot : 'button';
 
   return (
-    <Comp
+    <Comp 
       className={clsx(
-        'py-4 px-3 bg-cyan-500 rounded font-semibold text-black text-sm w-full transition-colors hover:bg-cyan-300 focus:ring-2 ring-white',
+        'py-3 px-4 bg-cyan-500 rounded font-semibold text-black text-sm w-full transition-colors hover:bg-cyan-300 focus:ring-2 ring-white',
+        className,
       )}
+      {...props}
     >
-      { children }
+      {children}
     </Comp>
   )
 }
