@@ -1,28 +1,27 @@
-import { Slot } from '@radix-ui/react-slot';
-import { clsx } from 'clsx';
 import { InputHTMLAttributes, ReactNode } from 'react';
+import { Slot } from '@radix-ui/react-slot';
 
-export interface ITextInputRootProps {
+export interface TextInputRootProps {
   children: ReactNode;
 }
 
-function TextInputRoot(props: ITextInputRootProps) {
+function TextInputRoot(props: TextInputRootProps) {
   return (
-    <div className='flex items-center gap-3 h-12 py-4 px-3 rounded bg-gray-800 focus-within:ring-2 ring-cyan-300'>      
-      { props.children }
+    <div className="flex items-center gap-3 h-12 py-4 px-3 rounded bg-gray-800 w-full focus-within:ring-2 ring-cyan-300">
+      {props.children}
     </div>
   )
 }
 
 TextInputRoot.displayName = 'TextInput.Root'
 
-export interface ITextInputIconProps {
+export interface TextInputIconProps {
   children: ReactNode;
 }
 
-function TextInputIcon(props: ITextInputIconProps) {
+function TextInputIcon(props: TextInputIconProps) {
   return (
-    <Slot className='w-6 h-6 text-gray-400'>
+    <Slot className="w-6 h-6 text-gray-400">
       {props.children}
     </Slot>
   )
@@ -30,13 +29,12 @@ function TextInputIcon(props: ITextInputIconProps) {
 
 TextInputIcon.displayName = 'TextInput.Icon'
 
-export interface ITextInputInputProps extends InputHTMLAttributes<HTMLInputElement> {}
+export interface TextInputInputProps extends InputHTMLAttributes<HTMLInputElement> {}
 
-function TextInputInput(props: ITextInputInputProps) {
-
+function TextInputInput(props: TextInputInputProps) {
   return (
-    <input
-      className='bg-transparent flex-1 outline-none text-gray-100 text-xs placeholder:text-gray-400'
+    <input 
+      className="bg-transparent flex-1 text-gray-100 text-xs placeholder:text-gray-400 outline-none"
       {...props}
     />
   )
@@ -44,9 +42,8 @@ function TextInputInput(props: ITextInputInputProps) {
 
 TextInputInput.displayName = 'TextInput.Input'
 
-
 export const TextInput = {
   Root: TextInputRoot,
   Input: TextInputInput,
-  Icon: TextInputIcon
+  Icon: TextInputIcon,
 }
