@@ -1,17 +1,18 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    './src/**/*.tsx'
+    './src/**/*.tsx',
   ],
   theme: {
     fontSize: {
       xs: 14,
-      s: 16,
+      sm: 16,
       md: 18,
       lg: 20,
       xl: 24,
       '2xl': 32,
     },
+
     colors: {
       transparent: 'transparent',
 
@@ -19,23 +20,22 @@ module.exports = {
       white: '#FFF',
 
       gray: {
-        100: '#E1E1E6',
-        200: '#c4c4cc',
-        400: '#7c7c8a',
-        800: '#202024',
         900: '#121214',
+        800: '#202024',
+        400: '#7c7c8a',
+        200: '#c4c4cc',
+        100: '#e1e1e6',
       },
 
       cyan: {
-        300: '#9be1fb',
-        500: '#81D8F7'
-      }
-      
+        500: '#81d8f7',
+        300: '#9BE1FB'
+      },
     },
     extend: {
       fontFamily: {
         sans: 'Inter, sans-serif'
-      }
+      },
     },
   },
   plugins: [],
